requirements.txt
asyncio
asyncpg
discord.py[voice, speed]
psutil
Wavelink
Pillow
spotipy
dblpy
psycopg2
pony